Resolve plugin dependencies for all projects
Resolve the plugin dependencies for all projects, including subprojects. This fixes the case where a subproject uses a plugin declaration that is resolved through a resolutionStrategy. In this case the versions update plugin will not report the Gradle plugin resolution artifact, but the actual artifact, just like when the plugin is added to the root build script using a classpath declaration. Because it's not actually added as a classpath dependency, the plugin needs to traverse the subprojects for buildscript dependencies too, otherwise the artifact won't be recognized as a plugin and would then be removed from the plugins block on updates.

Fixes #11