Filter transitive dependencies reported as current from versions report
If a version is reported as current, but the version is different from the version catalog, consider that version to be transitive, and remove it from the catalog that is generated from the versions report.

This fixes the issue where a version is downgraded because it's not used but declared in the catalog, possibly breaking version references etc in the process.