Setup publishing to the Gradle plugin portal
The plugin-publish plugin requires the shadow jar artifact to have no classifier (probably a good move anyway), switch to the maven publish base plugin so that no publications are added by default. The publication now seems to be configured by the plugin-publish plugin and the pom by the maven publish base plugin which is fine I suppose.