Skip redirected artifacts from version selector
Relax the null check for current version.
If we don't have a current version, then the resolving process was redirected to a different coordinate, that is not specified in the version catalog.
Before it is redirected though, the original coordinate is passed to the version selector, which is all we need and care about.